Split telemetry and tunables Python sources
